pyproject.toml
[project]
name = "bibleqandawebsite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=44.0.3",
    "django>=5.2",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.39.0",
    "google-auth-oauthlib>=1.2.2",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
]
prod = [
    "dj-database-url>=2.3.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.9.0",
]
